config controller for cross origin

+++ corejava/src/model/DynamicProgramming.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 public class DynamicProgramming {
 
-    private final Map<Integer, Integer> memo = new HashMap<>();
+    private final Map<Integer, Long> memo = new HashMap<>();
 
     public int secondLargest(int[] input) throws Exception {
         if (input.length > 1) {
@@ -38,7 +38,7 @@ public class DynamicProgramming {
      * Time Complexity: O(N)
      * Auxiliary Space: O(N)
      */
-    public int fibonacciTabulation(int n)
+    public long fibonacciTabulation(int n)
     {
         // Declare an array to store
         // Fibonacci numbers.
@@ -68,7 +68,7 @@ public class DynamicProgramming {
      * Time Complexity: O(N)
      * Auxiliary Space: O(N)
      */
-    public int fibonacciMemoization(int n)
+    public long fibonacciMemoization(int n)
     {
         if (n <= 1) {
             return n;
@@ -76,7 +76,7 @@ public class DynamicProgramming {
         if (memo.containsKey(n)) {
             return memo.get(n);
         }
-        int result = fibonacciMemoization(n - 1) + fibonacciMemoization(n - 2);
+        long result = fibonacciMemoization(n - 1) + fibonacciMemoization(n - 2);
         memo.put(n, result);
         return result;
     }
@@ -86,13 +86,13 @@ public class DynamicProgramming {
      * Time Complexity: O(N)
      * Auxiliary Space: O(1)
      */
-    public int fibonacciLoop(int n)
+    public long fibonacciLoop(int n)
     {
         if (n <= 1) {
             return n; // Base cases: fib(0) = 0, fib(1) = 1
         }
 
-        int f0 = 0, f1 = 1, f = 1;
+        long f0 = 0, f1 = 1, f = 1;
 
         for (int i = 2; i <= n; i++) {
             // Swap
@@ -108,7 +108,7 @@ public class DynamicProgramming {
      * Time Complexity: O(2**N)
      * Auxiliary Space: O(n)
      */
-    public int fibonacciRecursive(int n) {
+    public long fibonacciRecursive(int n) {
         if (n <= 1) {
             return n; // Base cases: fib(0) = 0, fib(1) = 1
         }
